+++ tests/python/pypackagecomplex/complex/binary_data.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-data = "\xf0\xf0\xf0"
+data = b"\xf0\xf0\xf0"